+++ command.py
@@ -66,7 +66,11 @@ class Command:
         return "cleaned"
 
     def reset(self) -> str:
-        shutil.copy(f'./memory/{self.Muice.user_qq}.json', './memory/chat_memory_backup.json')
+        if os.path.isfile(f'./memory/{self.Muice.user_qq}_backup.json'):
+            os.remove(f'./memory/{self.Muice.user_qq}_backup.json')
+        if not os.path.isfile(f'./memory/{self.Muice.user_qq}.json'):
+            return "你说得对，但是本雪和你没有说过一句话噢"
+        shutil.copy(f'./memory/{self.Muice.user_qq}.json', f'./memory/{self.Muice.user_qq}_backup.json')
         os.remove(f'./memory/{self.Muice.user_qq}.json')
         self.Muice.history = []
         return "reseted"
